# Conflicts:
#	src/test/java/org/gridsuite/modification/server/modifications/BalancesAdjustmentTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/BatteryCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/BatteryCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/BatteryModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/DeleteAttachingLineTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/DeleteVoltageLevelOnLineTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/EquipmentAttributeModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/EquipmentDeletionTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GenerationDispatchTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GeneratorCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GeneratorCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GeneratorModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GeneratorScalingTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/GroovyScriptTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LccCreationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineAttachToVoltageLevelTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineCreationInMixedTypologyTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineSplitWithNewVoltageLevelTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LineSplitWithVoltageLevelTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LinesAttachToSplitLinesTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LoadCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LoadCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LoadModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/LoadScalingTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/ModificationReferenceTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/ModificationUtilsTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/MoveVoltageLevelFeederBaysTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/OperatingStatusModificationEnergiseSideOneLineTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/OperatingStatusModificationEnergiseSideTwoLineTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/OperatingStatusModificationLockoutLineTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/ShuntCompensatorCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/ShuntCompensatorCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/ShuntCompensatorModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/StaticVarCompensatorCreationInBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/StaticVarCompensatorCreationInNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/SubstationModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/TwoWindingsTransformerCreationBusBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/TwoWindingsTransformerCreationMixedBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/TwoWindingsTransformerCreationNodeBreakerTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/TwoWindingsTransformerModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VoltageInitModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VoltageLevelCreationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VoltageLevelModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VoltageLevelTopologyModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VscCreationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/VscModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilter/assignment/AbstractModificationByAssignmentTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilter/assignment/TwoWindingsTransformerModificationByAssignmentTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilter/formula/AbstractByFormulaModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilter/formula/TwoWindingsTransformerByFormulaModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilter/formula/VoltageLevelByFormulaModificationTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/byfilterdeletion/AbstractByFilterDeletionTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/tabularmodifications/TabularGeneratorModificationsTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/tabularmodifications/TabularShuntCompensatorModificationsTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/tabularmodifications/TabularSubstationModificationsTest.java
#	src/test/java/org/gridsuite/modification/server/modifications/tabularmodifications/TabularVoltageLevelModificationsTest.java

+++ src/main/java/org/gridsuite/modification/server/entities/equipment/creation/LineSegmentEntity.java
@@ -36,7 +36,7 @@ public class LineSegmentEntity {
     private String segmentTypeId;
 
     @Column(name = "segment_distance_value")
-    private Integer segmentDistanceValue;
+    private Double segmentDistanceValue;
 
     @Column(name = "area")
     private String area;
